requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bottle==0.13.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.0.4
colorama==0.4.6
docker==5.0.3
FeatureCloud==0.0.29
flake8==3.9.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
joblib==1.4.2
jsonpickle==3.4.1
macholib==1.16.3
mccabe==0.6.1
multidict==6.1.0
neo4j==5.17.0
numpy==1.26.4
packaging==24.1
pandas==2.2.0
propcache==0.2.0
pycodestyle==2.7.0
pycparser==2.22
pydot==3.0.2
pyflakes==2.3.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.6
tzdata==2024.2
urllib3==1.26.20
websocket-client==1.8.0
yarl==1.17.1
